rename defaultmenu to template
