+++ src/main/java/fr/openmc/core/features/city/sub/mayor/menu/create/MayorColorMenu.java
@@ -2,7 +2,7 @@ package fr.openmc.core.features.city.sub.mayor.menu.create;
 
 import dev.lone.itemsadder.api.FontImages.FontImageWrapper;
 import fr.openmc.api.menulib.Menu;
-import fr.openmc.api.menulib.defaultmenu.ConfirmMenu;
+import fr.openmc.api.menulib.template.ConfirmMenu;
 import fr.openmc.api.menulib.utils.InventorySize;
 import fr.openmc.api.menulib.utils.ItemBuilder;
 import fr.openmc.core.features.city.City;
