+++ src/main/java/fr/openmc/core/features/city/sub/rank/CityRankAction.java
@@ -1,7 +1,7 @@
 package fr.openmc.core.features.city.sub.rank;
 
 import fr.openmc.api.input.DialogInput;
-import fr.openmc.api.menulib.defaultmenu.ConfirmMenu;
+import fr.openmc.api.menulib.template.ConfirmMenu;
 import fr.openmc.core.features.city.City;
 import fr.openmc.core.features.city.CityManager;
 import fr.openmc.core.features.city.CityPermission;
