+++ src/main/java/fr/openmc/core/features/city/sub/war/actions/WarActions.java
@@ -1,6 +1,6 @@
 package fr.openmc.core.features.city.sub.war.actions;
 
-import fr.openmc.api.menulib.defaultmenu.ConfirmMenu;
+import fr.openmc.api.menulib.template.ConfirmMenu;
 import fr.openmc.core.OMCPlugin;
 import fr.openmc.core.features.city.City;
 import fr.openmc.core.features.city.CityManager;
